Refactor workshop registrations page and components/dashboard directory.
- Updated column definitions to use `accessorFn` for participant details.
- Removed deprecated components..
