--- src/components/dashboards/checklist/types.ts
@@ -1,4 +1,0 @@
-export interface ChecklistState {
-  discordJoined: boolean;
-  devpostSignup: boolean;
-}
